% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_edge_nodelist.R
\name{align_edge_nodelist}
\alias{align_edge_nodelist}
\title{Copies the (normative) value of nodes from the nodelist to the edgelist.}
\usage{
align_edge_nodelist(edgelist, nodelist)
}
\arguments{
\item{edgelist}{an edgelist}

\item{nodelist}{an nodelist}
}
\value{
Returns an edgelist and nodelist
}
\description{
This function is used to align the values of concepts in the edgelist and
nodelist so that the value of the concepts in both lists align with
each other.You only need to run this function if you changed the default
value of any of the concepts in the nodelist after running the
prepare_lists_from_edgelist' function. You do not need to run it
when you only added information in the paradigm or instruments columns of
the nodelist
}
\details{
It takes the edgelist which was returns from the 'prepare_lists_from_edgelist'
function and the nodelist returned by the same function with the manually made
changes in the value column. Categorizing concepts as paradigmatic or as
instruments is optional, but required to run the functions paradigm_support
and instrument_support functions - see more instructions in the documentation
of these functions).
}
\examples{
\dontrun{
# Load the data
data("edgelist", "nodelist")

speaker_nodelist <- align_edge_nodelist(edgelist, nodelist)[[1]]
speaker_edgelist <- align_edge_nodelist(edgelist, nodelist)[[2]]

You may store the nodelist as csv and fill in the value, paradigms and
instruments column via excel by running the following code:
write.csv(df,file='/.../new_file.csv',fileEncoding = "UTF-8")

}
}
