% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_support.R
\name{calc_support_by_category_loop}
\alias{calc_support_by_category_loop}
\title{Instrument Support}
\usage{
calc_support_by_category_loop(node_calc, categories)
}
\arguments{
\item{node_calc}{An object of class "DataFrame". Description of parameter}

\item{second}{An object of class "vector". Description of parameter}
}
\value{
Returns an object of class "Dataframe". Description of what the function returns
}
\description{
Calculate the support for policy instruments
}
\examples{
categories <- c("Intergovernmental", "Supranational", "bla")
test_node_calc <- calc_support_by_category_loop(test_node_calc, categories)
}
