% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_degrees.R
\name{calculate_degrees}
\alias{calculate_degrees}
\title{calculate_degrees}
\usage{
calculate_degrees(edgelist, nodelist)
}
\arguments{
\item{edgelist}{The function takes an edgelist (dataframe) including all relations in a CM.
The data needs to following column structure. The columns with an * include
the meta-data and are optional. For other researchprojects, other types of
meta-data may be important.

\if{html}{\out{<div class="sourceCode">}}\preformatted{ * from: the 'cause'-node/concept id (corresponding to the id in the nodeslist)

 * to: the 'effect'-node/concept id (corresponding to the id in the nodeslist)

 * weight: the weight of the edge/relation (the number of times the relation
    is mentioned in the raw data (speech/text/survey))

 * map_id*: a unique id for the source (speech/text/respondent) the CM
    is derived from

 * map_date*: the date of the source (speech/text/survey)the CM is derived
    from
}\if{html}{\out{</div>}}}

\item{nodelist}{The function takes a nodelist (dataframe) including all nodes/concepts in the CM
The data needs to following column structure. The columns with an * include
the meta-data and are optional. For other researchprojects, other types of
meta-data may be important:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ * id: unique id (number) for the node/concept

 * node_name: node name/concept in words
}\if{html}{\out{</div>}}}
}
\value{
The function returns a dataframe entitled "node_calc" with all calculated values as
well as the original data.
For the function to return and store the following output a dataframe with the
values, you need to insert the following code below the function in your script:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  * node_calc_name_period <- calc_degrees_goW(edgelist, nodelist)
}\if{html}{\out{</div>}}
}
\description{
The calc_degrees-gow function is used to analyse cognitive map (CM) data.
It calculates various standard CM measures and draws a first CM graph.
You need to run this function first before you can run the more
complicated functions in this project: (calc_dims, draw_final_map, evaluation_step).
For more information about this type of analysis, see the reference publication
for this github repository (Van Esch & Snellens, forthcoming)
}
\details{
For each concept (node) in a map, the function calculates:

\if{html}{\out{<div class="sourceCode">}}\preformatted{* degree (also called centrality): the number of unique relations/edges
  by which the concept/node is connected to others (unweighted)

* indegree: the number of relations/edges feeding into a concept/node

* outdegree: the number of relations/edges feeding out of a concept/node

* the goal-orientation (go): (indegree-outdegree)/degree
}\if{html}{\out{</div>}}

Their weighted equivalents

\if{html}{\out{<div class="sourceCode">}}\preformatted{* weighted degree, also called saliency: the number of relations/edges
    by which the concept/node is connected to others taking into account
    the weight of the relations. This is calculated as the sum of the weights
    of all relations by which the concept is connnected

* indegree: the number of relations/edges feeding into a concept/node taking
    into account their weight (sum of weights of all ingoing relations)

* outdegree: the number of relations/edges feeding out of a concept/node
    taking into account their weight(sum of weights of all outgoing relations)

* the weighted goal-orientation (gow): (W indegree - W outdegree)/W degree
}\if{html}{\out{</div>}}
}
