% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cm.R
\name{draw_cm}
\alias{draw_cm}
\title{Displays the cognitive map as a simple graph.}
\usage{
draw_cm(edgelist, nodelist)
}
\arguments{
\item{edgelist}{an edgelist}

\item{nodelist}{an nodelist}
}
\value{
Returns a visual graph
}
\description{
This function is used to provide a first simple visualisation of the CM as
recorded in the edgelist and nodelist. The "interactive_CM_visualization"
function in this package provides a more sophisticated and easier to read graph
(see below). This function takes an edgelist and nodelist, you may use either
your original nodelist or the node_measures list that is returned by functions
you may have used earlier in the analysis (like the 'calculate_degrees" function).
}
\examples{
\dontrun{
# draw the CM graph
cognitivemapr::draw_cm(rutte_p2_edgelist, rutte_p2_nodelist)

# to store the cm as a png or pdf (other formats are possible), use the
following code:

# Save as PDF
grDevices::pdf(pdf_file)
cognitivemapr::draw_cm(rutte_p2_edgelist, rutte_p2_nodelist)
grDevices::dev.off()

# Save as PNG
grDevices::png(png_file, width = 800, height = 800)
cognitivemapr::draw_cm(rutte_p2_edgelist, rutte_p2_nodelist)
grDevices::dev.off()
#'
}

}
