% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_support.R
\name{instrument_support}
\alias{instrument_support}
\title{Instrument Support}
\usage{
instrument_support(node_measures, instruments)
}
\arguments{
\item{node_measures}{An object of class "dataframe" including all measures
returned when running the functions calculate_degrees & evaluate_concepts, as
well as a column with the title "instruments" in which the concepts are
classified as belonging to a pre-determined set of policy instrument type. If
a concept does not refer to a policy instrument, the cell should be left empty.}

\item{instruments}{An object of class "vector" containing all of the types of
policy instruments that occur in the node_measures list. To derive such a vector
run the following code:
"instruments <- unique(node_measures$instruments) #deriving all instrument-types
from the node_measures dataframe
instruments <- na.omit(instruments) #omitting the empty cells (NULL category)
from the analysis"}
}
\value{
Returns an object of class "dataframe" with a additional columns with
the instrument types as column titles and the saliency scores for those concepts
that are evaluated positively and that belong to these categories.
}
\description{
Calculates to what extent the cognitive map signals support for certain types
of policy instruments. More specifically, it determines the saliency of all
concepts classified as a type of policy instrument which are valued positively
as determined via the evaluate_concepts function. To be able to run this
function, the functions calculate-degrees AND evaluate_concepts  should have
been run first. Also the nodelist (node_measures) list this function takes,
needs to contain a column with the name 'instruments' in which for each concept
it is noted what type of policy instrument it represents. If a concept does
not refer to a policy instrument, the cell should be left empty.
The researcher can use any typology of policy instruments that is relevant to
their research.
}
\examples{
# deriving all instrument-typesfrom the node_measures dataframe
\dontrun{
library(readr)
load("./data/rutte_p2_edgelist.rda")
load("./data/rutte_p2_nodelist.rda")
instruments <- base::unique(rutte_p2_node_measures$instruments)
# omitting the empty cells (NULL category) from the analysis
instruments <- base::na.omit(instruments)
rutte_p2_node_measures <- cognitivemapr::calculate_degrees(rutte_p2_edgelist, rutte_p2_nodelist)
rutte_p2_node_measures <- instrument_support(rutte_p2_node_measures, instruments)
}
}
