% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_visual_cm.R
\name{interactive_visual_cm}
\alias{interactive_visual_cm}
\title{Creates an interactive graph of the CM}
\usage{
interactive_visual_cm(edgelist, node_measures)
}
\arguments{
\item{edgelist}{an edgelist prepared to work with VisNetworks}

\item{nodelist}{an node_measures prepared to work with VisNetworks}
}
\value{
Returns an VisNetwork interactive graph
}
\description{
It needs to be run on the edgelist and node_measures list and to have run the
calculate_degrees, evaluate_concepts, paradigm_support and instrument_support
prior to running this function to reap the function's full potential.
It uses the 'layout nicely' algorithm, you can move the concepts
zoom in and out, it shows the sign of the relations in colors, as well as
different node forms for the different paradigmatic concepts. it creates containers for
the different instruments so you can show the 'instrument-concepts individually
or collapse them into one category. The resulting CM also has legends.
}
\examples{
\dontrun{
#

To save the graph run: cognitive_map \%>\% visSave(file = "results/cognitive_map.html")
}
}
