% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cm_visualisation.R
\name{prep_cm_visualisation}
\alias{prep_cm_visualisation}
\title{Prepares the edge and node_measures list for VisNetwork graph}
\usage{
prep_cm_visualisation(edgelist, node_measures)
}
\arguments{
\item{edgelist}{an edgelist}

\item{nodelist}{an node_measures list}
}
\value{
Returns a list of lists
}
\description{
It needs to be run on the edgelist and node_measures list and to have run the
calculate_degrees, evaluate_concepts, paradigm_support and instrument_support
prior to running this function to reap the function's full potential.
It prepares the resulting egdelist and node_measure list for the interactive_cm
function
}
\examples{
\dontrun{
#


}
}
