% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutte_data.R
\docType{data}
\name{rutte_p2_node_measures}
\alias{rutte_p2_node_measures}
\title{Example nodelist: rutte_p2_nodelist}
\format{
A dataframe with 84 rows and 6 variables
\describe{
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
}
}
\usage{
rutte_p2_node_measures
}
\description{
A dataset ...
}
\references{
\url{data_blah.com}
}
\keyword{datasets}
