% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_edge_and_node_lists.R
\name{select_speaker_period_make_lists}
\alias{select_speaker_period_make_lists}
\title{select_speaker_period_make_lists}
\usage{
select_speaker_period_make_lists(name, start, eind)
}
\arguments{
\item{name}{Name of leader or actor}

\item{start}{start of period from which to draw the edges and nodes}

\item{eind}{end of period from which to draw the edges and nodes}
}
\value{
two data frames: one nodelist with the categorisations of the nodes
and intrinsic value (positive/negative concept) and a edgelist with the weight
of the edge and value (pos/neg/zero) and metadata
}
\description{
This is a helper function to select a particular set of edges & nodes from
the bigger edgelist \link{sumvalue_weighted_edges_leaders} and
nodelist \link{nodes_leaders_idecointval} that are part of the H2020 Transcrisis
research project of F.A.W.J. van Esch (Van Esch etal 2018). This database
will be made public at a later date.
You select the edges by (speaker)name and period and the function
returns the accompanying edgelist and nodelist.
Enter variables between " ", "eind" refers to the end date.
In using this function, you can name and store the created node and edgelist
reflecting its content by using the codes:
name_period_nodelist <- select_speaker_period_make_lists ("name", "start", "eind")[\link{1}]
name_period_edgelist <- select_speaker_period_make_lists ("name", "start", "eind")[\link{2}]
}
\examples{
x
}
