% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_iterations.R
\name{set_iterations}
\alias{set_iterations}
\title{Set the maximum number of iterations}
\usage{
set_iterations(edgelist, nodelist)
}
\arguments{
\item{edgelist}{an edgelist}

\item{nodelist}{a nodelist}
}
\value{
Returns a vector with the maximum number of iterations
}
\description{
This function is used to set the maximum number of iterations for the
evaluation of the nodes and edges in a CM. The maximum number of iterations
is equal to the diameter of the CM.
}
\examples{
set_iterations(rutte_p2_edgelist, rutte_p2_nodelist)
}
