% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cm.R
\name{draw_cm}
\alias{draw_cm}
\title{Displays the cognitive map as a simple graph.}
\usage{
draw_cm(edgelist, nodelist)
}
\arguments{
\item{edgelist}{an edgelist}

\item{nodelist}{an nodelist}
}
\value{
Returns a visual graph
}
\description{
This function is used to provide a first simple visualisation of the CM as
recorded in the edgelist and nodelist. The "interactive_CM_visualization
function in this package provides a more sufficated and easier to read graph
(see below).It takes an edgelist and nodelist, you may use either the original
nodelist or the node_measures list.
}
\examples{
\dontrun{
# draw the CM graph
cm_rutte_p2 <- draw_cm(rutte_p2_edgelist, rutte_p2_nodelist)

}
}
