% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_concepts.R
\name{evaluation_step}
\alias{evaluation_step}
\title{Conduct the evaluation analysis}
\usage{
evaluation_step(edgelist, nodelist)
}
\arguments{
\item{edgelist}{an edgelist}

\item{nodelist}{a nodelist}
}
\value{
Returns a list with the resulting edgelist and nodelist
}
\description{
This function is used to conduct the evaluation analysis of the nodes and
edges in a CM.
}
\examples{
# Load the data
data("edgelist")
data("nodelist")

# Run the evaluation analysis
max_runs <- set_iterations(edgelist, nodelist)
iterations <- length(max_runs)
result_list <- vector("list", iterations)

for (i in max_runs) {
   result_list <- Evaluation_step(edgelist, nodelist)
   edgelist <- result_list[[i]][[1]]
   nodelist <- result_list[[i]][[2]]
}

}
