% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalutation_step.R
\name{evaluation_step}
\alias{evaluation_step}
\title{Conduct an evaluation step}
\usage{
evaluation_step(edgelist, nodelist)
}
\arguments{
\item{edgelist}{an edgelist}

\item{nodelist}{a nodelist, if you want to add the evaluation
to the dataframe with the basic CM measures as calculated above, be
sure to use the 'node_measures' list that was returned when running the
calculate_degrees function.}
}
\value{
Returns a list with the resulting edgelist and nodelist
}
\description{
This function runs a single evaluation step
}
