% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cm_visualisation.R
\name{prep_cm_visualisation}
\alias{prep_cm_visualisation}
\title{Prepares the edge and node_measures list for VisNetwork graph}
\usage{
prep_cm_visualisation(edgelist, node_measures)
}
\arguments{
\item{edgelist}{an edgelist}

\item{node_measures}{an node_measures list}
}
\value{
Returns a list of lists
}
\description{
This function needs to be run on the edgelist and node_measures list and you
will need to have run the calculate_degrees, evaluate_concepts, paradigm_support
and instrument_support functions on your data prior to running this function
to reap the function's full potential.
This function reformats the egdelist and node_measure lists resulting from the
previously ran function so you can make create these into an interactive_cm
with the "interactive_visual_cm" function in this package which in turn relies
on the VisNetworks package.
}
\examples{
\dontrun{
#


}
}
