% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutte_p2_edgelist.R
\docType{data}
\name{rutte_p2_edgelist}
\alias{rutte_p2_edgelist}
\title{rutte_p2_edgelist}
\format{
A data frame with 106 rows and 10 variables:
\describe{
\item{from}{\emph{num} numerical id for the cause concept}
\item{to}{\emph{num} numerical id for the effect concept}
\item{weight}{\emph{num} number of times the relation is mentioned (can have any number)}
\item{edge_value}{\emph{num} sign/value of the relation: can be 1 (positive relation), -1 (negative relation), or 0 (non-existent relation)}
\item{edge_id}{\emph{num} unique numerical id for the edge/relation}
\item{map_id}{\emph{num} numerical id for this cognitive map}
\item{map_date}{\emph{Date} the date of the speech from which the relation was derived}
\item{speaker}{\emph{chr} name of the actor from whose assertions the edgelist was derived}
\item{value.x}{\emph{num} the evaluation (pos, neg ambiguous) of the 'from' concept}
\item{value.y}{\emph{num} the evaluation (pos, neg ambiguous) of the 'to' concept}
}
}
\source{
ain of ideational policy studies. Journal of European Public Policy, 31(2), 428–451. \url{https://doi.org/10.1080/13501763.2022.2155215}
}
\usage{
rutte_p2_edgelist
}
\description{
This dataset consists of an edgelist with causal and utility relations derived
from speeches by the Dutch prime-minister Mark Rutte on the Eurozone crisis
from the period from 2nd May 2010 to 26th July 2012.
}
\details{
The variables are as follows:
}
\keyword{datasets}
