% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutte_data.R
\docType{data}
\name{rutte_p2_edgelist}
\alias{rutte_p2_edgelist}
\title{rutte_p2_edgelist}
\format{
A dataframe with 106 rows and 10 variables
\describe{
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
\item{column}{description}
}
}
\usage{
rutte_p2_edgelist
}
\description{
A dataset ...
}
\references{
\url{data_blah.com}
}
\keyword{datasets}
