% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutte_p2_node_measures.R
\docType{data}
\name{rutte_p2_node_measures}
\alias{rutte_p2_node_measures}
\title{rutte_p2_node_measures}
\format{
A data frame with 84 rows and 14 variables:
\describe{
\item{id}{\emph{num} numerical id for all unique concepts in the associated edgelist}
\item{node_name}{\emph{chr} name of  all unique concepts in the associated edgelist}
\item{paradigms}{\emph{chr} the categorisation of the concept as belonging to/being associated with one of two incommensurable paradigms}
\item{int}{\emph{chr} the categorisation of the concept as intergovernmental or supranational (superfluous column for now)}
\item{value}{\emph{num} the intrinsic (normative) value of the concept displayed as a number: 1 = positive/neutral, -1 = negative.}
\item{instruments}{\emph{chr} the categorisation of the concept as a particular type of policy-instrument. This column can also be used for any other type of categorization of the concepts that
is deemed relevant}
\item{indegree}{indegree}
\item{outdegree}{outdegree}
\item{degree}{degree}
\item{w_indegree}{w_indegree}
\item{w_outdegree}{w_outdegree}
\item{w_degree}{w_degree}
\item{go}{go}
\item{gow}{gow}
}
}
\source{
van Esch, F., & Snellens, J. (2024). How to ‘measure’ ideas. Introducing the method of cognitive mapping to the domain of ideational policy studies. Journal of European Public Policy, 31(2), 428–451. \url{https://doi.org/10.1080/13501763.2022.2155215}
}
\usage{
rutte_p2_node_measures
}
\description{
This is an intermediate dataset used for examples in the vignette
}
\keyword{datasets}
