% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutte_p2_nodelist.R
\docType{data}
\name{rutte_p2_nodelist}
\alias{rutte_p2_nodelist}
\title{rutte_p2_nodelist}
\format{
A data frame with 84 rows and 6 variables:
\describe{
\item{id}{\emph{num} numerical id for all unique concepts in the associated edgelist}
\item{node_name}{\emph{chr} name of  all unique concepts in the associated edgelist}
\item{paradigms}{\emph{chr} the categorisation of the concept as belonging to/being associated with one of two incommensurable paradigms}
\item{int}{\emph{chr} the categorisation of the concept as intergovernmental or supranational (superfluous column for now)}
\item{value}{\emph{num} the intrinsic (normative) value of the concept displayed as a number: 1 = positive/neutral, -1 = negative.}
\item{instruments}{\emph{chr} the categorisation of the concept as a particular type of policy-instrument. This column can also be used for any other type of categorization of the concepts that
is deemed relevant}
}

A data frame with 84 rows and 6 variables:
\describe{
\item{id}{numerical id for all unique concepts in the associated edgelist}
\item{node_name}{name of  all unique concepts in the associated edgelist}
\item{paradigms}{the categorisation of the concept as belonging to/being associated with one of two incommensurable paradigms}
\item{int}{the categorisation of the concept as intergovernmental or supranational (superfluous column for now)}
\item{value}{the intrinsic (normative) value of the concept displayed as a number: 1 = positive/neutral, -1 = negative.}
\item{instruments}{the categorisation of the concept as a particular type of policy-instrument. This column can also be used for any other type of categorization of the concepts that
is deemed relevant}
}
}
\source{
van Esch, F., & Snellens, J. (2024). How to ‘measure’ ideas. Introducing the method of cognitive mapping to the domain of ideational policy studies. Journal of European Public Policy, 31(2), 428–451. \url{https://doi.org/10.1080/13501763.2022.2155215}
}
\source{
van Esch, F., & Snellens, J. (2024). How to ‘measure’ ideas. Introducing the method of cognitive mapping to the domain of ideational policy studies. Journal of European Public Policy, 31(2), 428–451. \url{https://doi.org/10.1080/13501763.2022.2155215}
}
\usage{
rutte_p2_nodelist

rutte_p2_nodelist
}
\description{
The nodelist containing all unique concepts in the edgelist with additional
information regarding these concepts, such as their intrinsic normative value
or their categorisation in terms of paradigms or (policy) instruments.
}
\details{
The data includes a nodelist with all of the concepts that are used in \emph{rutte_p2_edgelist},
categorised in terms of the policy-paradigms and policy-instruments that they
are associated with.

The variables are as follows:
}
\keyword{datasets}
