% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_iterations.R
\name{set_iterations}
\alias{set_iterations}
\title{Set Iterations
Helper function. Calculates the maximum number of iterations that the
evaluate_concepts function needs to run in the for-loop to arrive at the
correct evaluation of the concepts.
The maximum number of iterations is equal to the diameter of the CM.}
\usage{
set_iterations(edgelist, nodelist)
}
\arguments{
\item{edgelist}{an edgelist}

\item{nodelist}{a nodelist}
}
\value{
Returns a vector with the maximum number of iterations
}
\description{
Set Iterations
Helper function. Calculates the maximum number of iterations that the
evaluate_concepts function needs to run in the for-loop to arrive at the
correct evaluation of the concepts.
The maximum number of iterations is equal to the diameter of the CM.
}
\examples{
set_iterations(rutte_p2_edgelist, rutte_p2_nodelist)
}
