% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_iterations.R
\name{set_iterations}
\alias{set_iterations}
\title{Sets the number of iterations for the evaluate concepts function}
\usage{
set_iterations(edgelist, nodelist)
}
\arguments{
\item{edgelist}{an edgelist}

\item{nodelist}{a nodelist}
}
\value{
Returns a vector with the maximum number of iterations
}
\description{
Helper function. Calculates the number of iterations that the
evaluate_concepts function needs to run in the for-loop to determine the
(normative) value of the concepts.
The number of iterations is set to the diameter of the CM.
}
\examples{
set_iterations(rutte_p2_edgelist, rutte_p2_nodelist)
}
